dm: Fix FDT_ADDR_T_NONE definition on 64-bit machines
When devfdt_get_addr_name cannot find a node,
fdt_stringlist_search returns -FDT_ERR_NOTFOUND, or -1
Which resolves to 0xffffffffffffffff on 64-bit machines.

When FDT_ADDR_T_NONE is (-1U), or 0xffffffff, comparison with FDT_ERR_NOTFOUND
will fail.

This was observed and tested in drivers/net/ti/am65-cpsw-nuss.c

Reference Source:
- https://github.com/varigit/ti-u-boot/blob/ti-u-boot-2021.01_var01/scripts/dtc/libfdt/libfdt.h#L16-L17
- https://github.com/varigit/ti-u-boot/blob/ti-u-boot-2021.01_var01/scripts/dtc/libfdt/fdt_ro.c#L774
- https://github.com/varigit/ti-u-boot/blob/ti-u-boot-2021.01_var01/drivers/net/ti/am65-cpsw-nuss.c#L780
- https://github.com/varigit/ti-u-boot/blob/ti-u-boot-2021.01_var01/drivers/net/ti/am65-cpsw-nuss.c#L534

Derived from upstream: https://github.com/u-boot/u-boot/commit/9876ae7db6da1cf8e106fcb46a36172fbb5781e5

+++ include/fdtdec.h
@@ -25,7 +25,7 @@ typedef phys_addr_t fdt_addr_t;
 typedef phys_size_t fdt_size_t;
 
 #ifdef CONFIG_PHYS_64BIT
-#define FDT_ADDR_T_NONE (-1U)
+#define FDT_ADDR_T_NONE ((ulong)(-1))
 #define fdt_addr_to_cpu(reg) be64_to_cpu(reg)
 #define fdt_size_to_cpu(reg) be64_to_cpu(reg)
 #define cpu_to_fdt_addr(reg) cpu_to_be64(reg)
